pyproject.toml
[tool.poetry]
name = "endpoint"
version = "0.1.0"
description = ""
authors = ["abdelfattah27 <Abdelfattah.hamdy234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
